// plugins/cad/src/types/Capacity.ts
/**
 * Copyright 2024 The Nephio Authors.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { KubernetesKeyValueObject } from './KubernetesResource';

export type Capacity = {
  readonly apiVersion: string;
  readonly kind: string;
  readonly metadata: CapacityMetadata;
  readonly spec: CapacitySpec;
};

export type CapacityMetadata = {
  readonly name: string;
  readonly namespace?: string;
  readonly labels?: KubernetesKeyValueObject;
  readonly annotations?: KubernetesKeyValueObject;
};

export type CapacitySpec = {
  readonly maxDownlinkThroughput?: string;
  readonly maxUplinkThroughput?: string;
  readonly maxNFConnections?: number;
  readonly maxSessions?: number;
  readonly maxSubscribers?: number;
};